Upgrade @openzeppelin/contracts
Upgraded  @openzeppelin/contracts dependency 3.0.0 -> 3.4.0. This
upgrade uncluded a breaking change, but it doesn't impact the
arbitrager example.

The breaking change is:
ERC20Snapshot: switched to using _beforeTokenTransfer hook instead of
overriding ERC20 operations.

This is the latest version of @openzeppelin/contracts that supports
Solidity compiler 0.6.x and should be upgraded once the target compiler
version is bumped to the latest version.